% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem_smiles.R
\docType{data}
\name{chem_smiles}
\alias{chem_smiles}
\title{Example Compounds and Associated SMILES.}
\format{
A data frame with 42 rows and 2 variables:
\describe{
\item{compound_id}{Descriptive names of forty-two chemical compounds.}
\item{smile}{The corresponding SMILE of the compounds.}
}
}
\usage{
chem_smiles
}
\description{
A dataset providing example simplified molecular-input line-entry system ("SMILES") strings for use in chemdesr.
}
\keyword{datasets}
