% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdesr-package.R
\docType{package}
\name{chemdesr-package}
\alias{chemdesr}
\alias{chemdesr-package}
\title{chemdesr: Obtain Data from the ChemDes Platform.}
\description{
This package scrapes the ChemDes platform for chemical information.
}
\author{
\strong{Maintainer}: Jack Davison \email{jd1184@york.ac.uk} (\href{https://orcid.org/0000-0003-2653-6615}{ORCID})

}
\keyword{internal}
