% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_chemdes.R
\name{map_chemdes}
\alias{map_chemdes}
\title{Map ChemDes}
\usage{
map_chemdes(smiles, chatty = F)
}
\arguments{
\item{smiles}{A vector of simplified molecular-input line-entry system ("SMILES") strings.}

\item{chatty}{TRUE/FALSE. If TRUE, will print messages to the console updating you of the stage of scraping.}
}
\value{
A "wide" tibble of chemical information, labelled with the SMILES string.
}
\description{
This function is a wrapper around read_full_chemdes, allowing for a vector of SMILES strings to be mapped over. This function should take roughly 30 seconds per compound.
}
\author{
Jack Davison, \email{jd1184@york.ac.uk}
}
