% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chemdes.R
\name{read_chemdes}
\alias{read_chemdes}
\title{Read ChemDes}
\usage{
read_chemdes(smile, desc = "chemopy")
}
\arguments{
\item{smile}{A simplified molecular-input line-entry system ("SMILES") string.}

\item{desc}{The descriptors you wish to scrape. One of "Chemopy", "CDK", "RDKit", "Pybel", or "PaDEL" (not case sensitive).}
}
\value{
A "wide" tibble of chemical information, labelled with the SMILES string.
}
\description{
This function scrapes chemical information from ChemDes, a free web-based platform for the calculation of molecular descriptors and fingerprints.
}
\author{
Jack Davison, \email{jd1184@york.ac.uk}
}
